[#40] Rename missing packages
