+++ src/main/java/com/github/ruromero/cloudeventsplayer/model/PlayerMode.java
@@ -1,4 +1,4 @@
-package com.redhat.syseng.tools.cloudevents.model;
+package com.github.ruromero.cloudeventsplayer.model;
 
 public enum PlayerMode {
     LOCAL,
