+++ src/main/java/com/github/ruromero/cloudeventsplayer/resources/MessagesSocket.java
@@ -1,4 +1,4 @@
-package com.redhat.syseng.tools.cloudevents.resources;
+package com.github.ruromero.cloudeventsplayer.resources;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
